test(studio): add comprehensive unit tests for Studio module
- Add 62 new unit test files covering Studio subsystems
- Tests cover commands, controllers, server, collectors, storage, evidence
- Fix EvidenceArchive::fromJson() to throw InvalidArgumentException
- Fix ConsoleVerifyCommand error handling for invalid archives
- Improve test coverage for CI threshold compliance

+++ tests/Unit/Studio/Console/Storage/SqliteEventStoreTest.php
@@ -873,7 +873,7 @@ final class SqliteEventStoreTest extends TestCase
             eventId: $eventId,
             eventType: $eventType,
             schemaVersion: EventVersion::V1,
-            timestampUs: $timestampUs ?? (int) ((float) microtime(true) * 1_000_000.0),
+            timestampUs: $timestampUs ?? (int) (microtime(true) * 1_000_000.0),
             requestId: $requestId,
             traceId: $traceId,
             spanId: 'span-123',
